% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpost.R
\name{getpost}
\alias{getpost}
\title{use GET to get a single post from support.bioconductor.org API}
\usage{
getpost(uid)
}
\arguments{
\item{uid}{a character(1) identifier 'uid' in the support API}
}
\description{
use GET to get a single post from support.bioconductor.org API
}
\examples{
mp = getpost("5905")
cmp = httr::content(mp)
names(cmp)
cmp$title
}
